Refactor list view for repacks
