+++ client/packages/common/src/intl/locales/en/common.json
@@ -155,6 +155,7 @@
   "label.message": "Message",
   "label.reason": "Reason",
   "label.log": "Log",
+  "label.split-move": "Split/Move",
   "label.manufacturer": "Manufacturer",
   "label.mode": "Mode",
   "label.months-of-stock": "MOS",
