+++ client/packages/system/src/Stock/Components/Repack/index.ts
@@ -1,1 +1,1 @@
-export * from './Repack';
+export * from './RepackView';
